# Optimize peekTQueue and peekTBQueue