Reduce the amount of operations, avoiding redundant writes and hence reducing the chance of conflicts