@nikita-volkov 2018-09-18
Unfortunately I don't have the time for that. Also I don't think that it needs proving that avoiding redundant writes reduces contention, and hence increases performance.