@mitchellwrosen 2018-09-18
LGTM, and I agree benchmarking this is not necessary. I did a quick and dirty one anyway:

- 2 threads each writing 2M elements
- 4 threads each peeking 2M elements

Old:

```
benchmarking ./peekbench +RTS -N4
time                 2.592 s    (2.285 s .. 3.051 s)
                     0.995 R²   (0.994 R² .. 1.000 R²)
mean                 2.516 s    (2.465 s .. 2.615 s)
std dev              85.33 ms   (0.0 s .. 86.57 ms)
variance introduced by outliers: 19% (moderately inflated)
```

New:

```
benchmarking ./peekbench +RTS -N4
time                 1.425 s    (1.215 s .. 1.575 s)
                     0.998 R²   (0.992 R² .. 1.000 R²)
mean                 1.408 s    (1.360 s .. 1.432 s)
std dev              41.60 ms   (0.0 s .. 41.77 ms)
variance introduced by outliers: 19% (moderately inflated)
```

Code:

```haskell
import Control.Concurrent
import Control.Concurrent.STM
import Control.Monad

main :: IO ()
main = do
  q <- newTQueueIO
  done <- newEmptyMVar

  forkIO $ replicateM_ 2000000 (atomically (writeTQueue q 1)) >> putMVar done ()
  forkIO $ replicateM_ 2000000 (atomically (writeTQueue q 1)) >> putMVar done ()

  forkIO $ replicateM_ 2000000 (atomically (peekTQueue q))    >> putMVar done ()
  forkIO $ replicateM_ 2000000 (atomically (peekTQueue q))    >> putMVar done ()
  forkIO $ replicateM_ 2000000 (atomically (peekTQueue q))    >> putMVar done ()
  forkIO $ replicateM_ 2000000 (atomically (peekTQueue q))    >> putMVar done ()

  replicateM_ 6 $ takeMVar done
```