@Fuuzetsu 2018-09-19
That's a huge improvement. Why not add these benchmarks into the package?