@mitchellwrosen 2018-09-19
It definitely could. Who should make that call? Not me :) Again I agree with @nikita-volkov that his implementation is just obviously better (for lack of a better word - I know it's not a good one to use when discussing code).

A peek is just a read that is _not_ followed by writing back the tail, and now it's implemented as such. So an understanding of how STM works, or really just an understanding of this patch, is (to my mind) sufficient proof that the old way (read + write + read + write) is worse.